Add airbnb eslint config + fix linting errors
