+++ helpers/utilityFunctions.ts
@@ -1,4 +1,4 @@
-import {IFailResponse} from '../interfaces/api'
+import { IFailResponse } from '../interfaces/api';
 
 export {};
 
@@ -6,9 +6,9 @@ export {};
    * helper fn for returning an error
    * @param message description of what went wrong
    */
-export function failReturn(message: string): IFailResponse {
+export default function failReturn(message: string): IFailResponse {
   return ({
     success: false,
     message,
-  })
+  });
 }
