--- interfaces/api.d.ts
@@ -1,15 +1,0 @@
-export interface IApiResponseBase {
-    success: boolean;
-    message: string;
-  }
-
-  export interface ISuccessResponse<T> extends IApiResponseBase {
-    success: true;
-    data: T;
-  }
-
-  export interface IFailResponse extends IApiResponseBase {
-    success: false;
-  }
-
-  export type IApiResponse<T> = ISuccessResponse<T> | IFailResponse
