+++ interfaces/api.ts
@@ -1,0 +1,15 @@
+export interface IApiResponseBase {
+    success: boolean;
+    message: string;
+  }
+
+export interface ISuccessResponse<T> extends IApiResponseBase {
+    success: true;
+    data: T;
+  }
+
+export interface IFailResponse extends IApiResponseBase {
+    success: false;
+  }
+
+export type IApiResponse<T> = ISuccessResponse<T> | IFailResponse
