+++ package.json
@@ -11,7 +11,7 @@
   "dependencies": {
     "autoprefixer": "^10.1.0",
     "classnames": "^2.2.6",
-    "eslint": "^7.16.0",
+    "eslint": "7.2.0",
     "http-status-codes": "^2.1.4",
     "lodash": "^4.17.20",
     "next": "^10.0.4",
@@ -23,13 +23,18 @@
     "tailwindcss": "^2.0.2"
   },
   "devDependencies": {
+    "@types/classnames": "^2.2.11",
+    "@types/lodash": "^4.14.166",
     "@types/node": "^12.12.21",
     "@types/react": "^16.9.16",
     "@types/react-dom": "^16.9.4",
-    "@types/classnames": "^2.2.11",
-    "@types/lodash": "^4.14.166",
     "@typescript-eslint/eslint-plugin": "^4.11.0",
     "@typescript-eslint/parser": "^4.11.0",
+    "eslint-config-airbnb": "18.2.1",
+    "eslint-plugin-import": "^2.22.1",
+    "eslint-plugin-jsx-a11y": "^6.4.1",
+    "eslint-plugin-react": "^7.21.5",
+    "eslint-plugin-react-hooks": "4.0.0",
     "typescript": "4.0"
   },
   "license": "MIT"
