+++ pages/api/greet/[name]/[count].ts
@@ -1,10 +1,8 @@
-
-
-import { StatusCodes } from 'http-status-codes'
-import { NextApiRequest, NextApiResponse } from 'next'
-import { repeat } from 'lodash'
-import { failReturn } from '../../../../helpers/utilityFunctions'
-import { IApiResponse } from '../../../../interfaces/api'
+import { StatusCodes } from 'http-status-codes';
+import { NextApiRequest, NextApiResponse } from 'next';
+import { repeat } from 'lodash';
+import { failReturn } from '../../../../helpers/utilityFunctions';
+import { IApiResponse } from '../../../../interfaces/api';
 
 const handler = (_req: NextApiRequest, res: NextApiResponse<IApiResponse<string>>): void => {
   const {
@@ -13,21 +11,21 @@ const handler = (_req: NextApiRequest, res: NextApiResponse<IApiResponse<string>
 
   try {
     if (typeof name !== 'string') {
-      throw new Error('Please specify a name as text')
+      throw new Error('Please specify a name as text');
     }
 
-    const repeatCount = parseInt(count as string, 10)
+    const repeatCount = parseInt(count as string, 10);
 
     const response = {
       success: true,
       message: `Successfully generated greeting for ${name}`,
-      data: repeat(`Hello, ${name}. `, repeatCount)
-    }
+      data: repeat(`Hello, ${name}. `, repeatCount),
+    };
 
-    res.status(StatusCodes.OK).json(response)
+    res.status(StatusCodes.OK).json(response);
   } catch (err) {
-    res.status(StatusCodes.INTERNAL_SERVER_ERROR).json(failReturn(`Something went wrong! crying_cat_face.`))
+    res.status(StatusCodes.INTERNAL_SERVER_ERROR).json(failReturn('Something went wrong! crying_cat_face.'));
   }
-}
+};
 
-export default handler
+export default handler;
