+++ pages/api/users/index.ts
@@ -1,27 +1,26 @@
-import { NextApiRequest, NextApiResponse } from 'next'
-import StatusCodes from 'http-status-codes'
-import { failReturn } from '../../../helpers/utilityFunctions'
-import { User } from '../../../interfaces'
-import { IApiResponse } from '../../../interfaces/api'
-import { sampleUserData } from '../../../utils/sample-data'
-
+import { NextApiRequest, NextApiResponse } from 'next';
+import StatusCodes from 'http-status-codes';
+import { failReturn } from '../../../helpers/utilityFunctions';
+import { User } from '../../../interfaces';
+import { IApiResponse } from '../../../interfaces/api';
+import { sampleUserData } from '../../../utils/sample-data';
 
 const handler = (_req: NextApiRequest, res: NextApiResponse<IApiResponse<User[]>>): void => {
   try {
     if (!Array.isArray(sampleUserData)) {
-      throw new Error('Cannot find user data')
+      throw new Error('Cannot find user data');
     }
 
     const response = {
       success: true,
       message: `retrieved ${sampleUserData.length} users`,
-      data: sampleUserData
-    }
+      data: sampleUserData,
+    };
 
-    res.status(StatusCodes.OK).json(response)
+    res.status(StatusCodes.OK).json(response);
   } catch (err) {
-    res.status(StatusCodes.INTERNAL_SERVER_ERROR).json(failReturn(`Unable to return users`))
+    res.status(StatusCodes.INTERNAL_SERVER_ERROR).json(failReturn('Unable to return users'));
   }
-}
+};
 
-export default handler
+export default handler;
